+++ resources/attributes/default.rb
@@ -65,6 +65,10 @@ default['cape']['min_freespace'] = 15000
 # drill
 default['drill']['port'] = 8047
 
+# redborder-hub
+default['redborder']['redborder-hub']['registered'] = false
+default['redborder']['redborder-hub']['port'] = 8010
+
 # hard disk
 default['redborder']['manager']['data_dev'] = {}
 default['redborder']['manager']['data_dev']['root'] = '/dev/mapper/VolGroup-lv_root'
@@ -111,6 +115,8 @@ default['redborder']['memory_services']['airflow-webserver'] = { 'count': 50, 'm
 default['redborder']['memory_services']['airflow-triggerer'] = { 'count': 10, 'memory': 0 }
 default['redborder']['memory_services']['airflow-dag-processor'] = { 'count': 20, 'memory': 0 }
 default['redborder']['memory_services']['rb-reputation'] = { 'count': 30, 'memory': 0 }
+default['redborder']['memory_services']['redborder-hub'] = { 'count': 10, 'memory': 0 }
+default['redborder']['memory_services']['logstash'] = { 'count': 80, 'memory': 0 }
 
 # default attributes for managers_info, it would be rewriten with the cluster config
 default['redborder']['cluster_info'] = {}
@@ -129,7 +135,7 @@ default['redborder']['services_group']['full'] = %w(consul chef-server zookeeper
                                                     druid-broker druid-historical druid-coordinator druid-router druid-indexer druid-overlord
                                                     postgresql nginx webui rb-workers f2k rb-druid-indexer
                                                     redborder-monitor sfacctd redborder-dswatcher redis
-                                                    redborder-events-counter http2k redborder-mem2incident rb-logstatter)
+                                                    redborder-events-counter http2k redborder-mem2incident rb-logstatter redborder-alarm-engine)
 
 default['redborder']['services_group']['custom'] = %w(consul)
 default['redborder']['services_group']['core'] = %w(consul chef-server s3 postgresql nginx)
@@ -139,7 +145,7 @@ default['redborder']['services_group']['historical'] = %w(consul druid-historica
 default['redborder']['services_group']['middlemanager'] = %w(consul druid-middlemanager)
 default['redborder']['services_group']['broker'] = %w(consul druid-broker)
 default['redborder']['services_group']['http2k'] = %w(consul http2k)
-default['redborder']['services_group']['webui'] = %w(consul nginx webui rb-workers)
+default['redborder']['services_group']['webui'] = %w(consul nginx webui rb-workers redborder-alarm-engine)
 default['redborder']['services_group']['f2k'] = %w(consul f2k)
 default['redborder']['services_group']['s3'] = %w(consul nginx s3)
 default['redborder']['services_group']['postgresql'] = %w(consul postgresql)
@@ -203,6 +209,8 @@ default['redborder']['services']['secor']                     = false
 default['redborder']['services']['secor-vault']               = false
 default['redborder']['services']['redis']                     = false
 default['redborder']['services']['rb-reputation']             = false
+default['redborder']['services']['redborder-hub']             = true
+default['redborder']['services']['redborder-alarm-engine']    = true
 
 default['redborder']['systemdservices']['airflow-scheduler']        = ['airflow-scheduler']
 default['redborder']['systemdservices']['airflow-webserver']        = ['airflow-webserver']
@@ -262,6 +270,8 @@ default['redborder']['systemdservices']['secor']                    = ['rb-secor
 default['redborder']['systemdservices']['secor-vault']              = ['rb-secor-vault']
 default['redborder']['systemdservices']['redis']                    = ['redis']
 default['redborder']['systemdservices']['rb-reputation']            = ['rb-reputation']
+default['redborder']['systemdservices']['redborder-hub']            = ['redborder-hub']
+default['redborder']['systemdservices']['redborder-alarm-engine']   = ['redborder-alarm-engine']
 
 # Balanced services
 default['redborder']['manager']['balanced'] = [