+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 cookbook-rb-manager CHANGELOG
 ===============
 
+## 5.29.0
+
+  - vimesa
+    - [6f28867] Add redborder-hub settings
+
+## 5.28.1
+
+  - Nils
+    - [c65c263] add logstash to cgroup
+
+## 5.28.0
+
+  - David Vanhoucke
+    - [8eb3050] add redborder-alarm-engine
+
 ## 5.27.3
 
   - Vimesa